+++ backend/src/main/java/kernel360/techpick/core/model/user/Role.java
@@ -1,4 +1,4 @@
-package kernel360.techpick.core.feature.user;
+package kernel360.techpick.core.model.user;
 
 public enum Role {
 	ROLE_GUEST, // 로그인 안한 사용자
